# index, restic: Genericize indexMap