… internal/index/indexmap.go
… internal/index/indexmap_test.go