@MichaelEischer 2022-12-16
> It should be possible to reimplement restic.CountedBlobSet using this type as well.

I like the idea in general, although I wonder how it would be possible to embed the counter byte directly into the normal repository index. That way we could minimize the memory overhead of the BlobSet to just a few percent of the repository index.

The index could the return a CountedBlobSet (only once, to prevent multiple users from conflicting) which can store that information directly in the index.