@greatroar 2023-01-27
I'm a bit closer to finding out what's going on with the performance of the "unknown" lookups: it's not the generics, because when I apply the same refactor without generics to indexMap, I get the same slowdown.