@MichaelEischer 2023-04-07
I guess the last commit fixes the performance problems? I'm not entirely happy with adding more code to `internal/restic`, but it's probably the best place for now.