@MichaelEischer 2023-06-08
My current impression is that https://github.com/restic/restic/pull/4354 is more likely a better solution than a generic indexMap. In particular the `AssociatedData` type in #4354 only requires about 1/70 the memory which would be necessary when using a generic indexMap.