@MichaelEischer 2023-07-09
After thinking a bit more about this PR, I still think that #4354 is the way forward. Thus, I'll close the PR for now.